pyproject.toml
[project]
name = "custom_plugins"
version = "0.1.0"
description = "Default template for PDM package"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "reine-ishyanami", email = "2402979195@qq.com" }]
dependencies = [
    "nonebot2[fastapi]>=2.3.1",
    "nonebot-adapter-onebot>=2.4.3",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-waiter>=0.6.2",
    "nonebot-plugin-localstore>=0.7.0",
    "aiofiles>=24.1.0",
    "nonebot-plugin-orm[sqlite]>=0.7.5",
    "playwright>=1.47.0",
    "nonebot-plugin-alconna>=0.53.0",
    "nonebot-plugin-uninfo>=0.5.0",
    "PyYAML>=6.0.2",
    "nonebot-plugin-htmlrender>=0.3.5",
    "ruamel-yaml>=0.18.6",
    "pillow>=11.1.0",
    "google-genai>=0.4.0",
]
license = { text = "MIT" }

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = [
    "nonebot_plugin_orm",
    "nonebot_plugin_alconna",
    "nonebot_plugin_uninfo",
    "nonebot_plugin_htmlrender",
]
plugin_dirs = ["forward_plugins", "direct_plugins", "no_deploy"]
builtin_plugins = ["echo"]

[tool.pdm]
distribution = false

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.ruff]
line-length = 120
ignore = ["F401", "E402"]
